Abstracts out the file input to hopefully limit breakage
